Add initial version of CMakeLists.txt files
This commit introduced 2 CMakeLists.txt files. The main CMakeLists.txt defines the global options of the project and defined the jpeg library. The second CMakeLists.txt defines the command line tool that uses the library.
The main 3 C++ compilers: MSVC, CGG and clang are supported).

Note: the shared library on Linux builds require a source code update to ensure not all functions are exported by default.